backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.2
django-environ>=0.11
django-import-export>=4.1
psycopg2-binary>=2.9
Pillow>=10.3
stripe>=10.0
openpyxl>=3.1
weasyprint>=62.0
gunicorn>=22.0
requests>=2.32
Faker>=24.0
celery>=5.3
redis>=5.0
